Reorganize paypal app
