+++ paypal/constants.py
@@ -1,0 +1,9 @@
+from utils import CurrencyCode
+
+DEFAULT_CURRENCY_CODE = CurrencyCode.USD
+
+PAYPAL_API_URL = "https://api.sandbox.paypal.com"
+PAYPAL_ORDER_BASE_URL = f"{PAYPAL_API_URL}/v2/checkout/orders"
+PAYPAL_SUBSCRIPTIONS_BASE_URL = f"{PAYPAL_API_URL}/v1/billing/subscriptions"
+
+ONE_DAY_S = 60 * 60 * 24  # 24 hours
