+++ paypal/tests.py
@@ -2,14 +2,19 @@ from unittest import mock
 from django.test import TestCase
 from requests.exceptions import HTTPError
 
-from .paypal import (
+
+from .orders import (
+    create_order,
+    capture_order,
+)
+from .subscriptions import (
+    get_subscription,
+    subscription_is_active,
+)
+from .utils import (
     PayPalError,
     get_auth_token,
     construct_paypal_auth_headers,
-    create_order,
-    capture_order,
-    get_subscription,
-    subscription_is_active,
 )
 
 
@@ -26,7 +31,7 @@ class GetAuthTokenTest(TestCase):
         result = get_auth_token()
         self.assertEqual(result, "some_token")
 
-    @mock.patch("paypal.paypal.requests.post")  # Replace 'your_module' accordingly
+    @mock.patch("paypal.paypal.requests.post")
     def test_get_auth_token_failure(self, mock_post):
         # Mock failed API response
         mock_response = mock.Mock()
@@ -35,5 +40,5 @@ class GetAuthTokenTest(TestCase):
         mock_post.return_value = mock_response
 
         # Test function should raise an error
-        with self.assertRaises(PayPalError):  # Replace with your actual exception class
+        with self.assertRaises(PayPalError):
             get_auth_token()
